Django_apps/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout>=4.0.2,<5.0
attrs==23.1.0
bokeh==3.2.2
cbor2==5.4.6
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
Django==4.2.5
django-debug-toolbar==4.2.0
django-filter==23.5
django-pivot==1.10.0
dnspython==2.4.1
frozenlist==1.3.3
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mssql-django==1.3
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
psycopg2-binary==2.9.9
pubnub==7.2.0
py4j==0.10.9.7
pycryptodomex==3.18.0
pymongo==4.4.1
pyodbc==5.0.1
pyspark==3.4.1
python-dateutil==2.8.2
python3-memcached==1.51
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.24
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.3
xyzservices==2023.5.0
yarl==1.9.2
xhtml2pdf
django-sesame
face_recognition
dbfread
oauth2client==4.1.3
django_descope
celery==5.3.6
redis==5.0.2
